+++ cli/azd/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Release History
 
-## 0.1.0-beta.3 (Unreleased)
+## 0.1.0-beta.4 (Unreleased)
+
+## 0.1.0-beta.3 (2022-07-28)
 
 ### Features Added
 
 - [[#100]](https://github.com/Azure/azure-dev/pull/100) Add support for an optional `docker` section in service configuration to control advanced docker options.
+- [[#152]](https://github.com/Azure/azure-dev/pull/152) While provisioning in interactive mode (default), Azure resources are now logged to console as they are created.
 
 ### Breaking Changes
 
@@ -16,6 +19,7 @@
 - [[#115]](https://github.com/Azure/azure-dev/issues/115) Fix deploy error when using a resource name with capital letters.
 
 ### Other Changes
+- [[#188]](https://github.com/Azure/azure-dev/issues/188) Update the minimum Bicep version to `v0.8.9`.
 
 ## 0.1.0-beta.2 (2022-07-13)